[POAE7-1420] code refactor to fix targetExpr info lost (#17)
* [POAE7-1420] code refactor to fix targetExpr info lost

* move exe unit get method in CiderParamContext
