+++ velox/core/HybridPlanNode.h
@@ -15,9 +15,8 @@
  */
 #pragma once
 
-#include "velox/core/PlanNode.h"
-
 #include "QueryEngine/RelAlgExecutionUnit.h"
+#include "velox/core/PlanNode.h"
 
 namespace facebook::velox::core {
 /**
@@ -31,18 +30,94 @@ struct NodeProperty {
   bool hasOrderBy;
 };
 
+struct CiderParamContext {
+  CiderParamContext(RowTypePtr rowType, PlanNodeId id)
+      : rowType_(rowType),
+        id_(id),
+        inputDescs_{},
+        inputColDescs_{},
+        simpleQuals_{},
+        quals_{},
+        targetExprMap_{},
+        groupByExprMap_{},
+        orderByCollation_{},
+        limit_(0),
+        offset_(0),
+        nodeProperty_({false, false, false, false, false}) {}
+
+  RowTypePtr rowType_;
+  PlanNodeId id_;
+  std::vector<InputDescriptor> inputDescs_;
+  std::list<std::shared_ptr<const InputColDescriptor>> inputColDescs_;
+  std::list<std::shared_ptr<Analyzer::Expr>> simpleQuals_;
+  std::list<std::shared_ptr<Analyzer::Expr>> quals_;
+  std::vector<std::pair<std::string, std::shared_ptr<Analyzer::Expr>>>
+      targetExprMap_;
+  std::vector<std::pair<std::string, std::shared_ptr<Analyzer::Expr>>>
+      groupByExprMap_;
+  std::list<Analyzer::OrderEntry> orderByCollation_;
+  size_t limit_;
+  size_t offset_;
+  NodeProperty nodeProperty_;
+  SortInfo getSortInfoFromCtx() {
+    if (orderByCollation_.size() == 0) {
+      return {{}, SortAlgorithm::Default, 0, 0};
+    } else {
+      return {
+          orderByCollation_,
+          SortAlgorithm::SpeculativeTopN,
+          limit_,
+          offset_};
+    }
+  }
+  std::shared_ptr<RelAlgExecutionUnit> getExeUnitBasedOnContext() {
+    std::vector<Analyzer::Expr*> targetExprs;
+    std::list<std::shared_ptr<Analyzer::Expr>> groupByExprs;
+    for (auto it : targetExprMap_) {
+      targetExprs.push_back(it.second.get());
+    }
+    if (targetExprs.size() == 0) {
+      return nullptr;
+    }
+    for (auto it : groupByExprMap_) {
+      groupByExprs.push_back(it.second);
+    }
+    if (groupByExprs.size() == 0) {
+      std::shared_ptr<Analyzer::Expr> empty;
+      groupByExprs.emplace_back(empty);
+    }
+    JoinQualsPerNestingLevel leftDeepJoinQuals;
+    return std::make_shared<RelAlgExecutionUnit>(RelAlgExecutionUnit{
+        inputDescs_,
+        inputColDescs_,
+        simpleQuals_,
+        quals_,
+        leftDeepJoinQuals,
+        groupByExprs,
+        targetExprs,
+        nullptr,
+        getSortInfoFromCtx(),
+        0,
+        RegisteredQueryHint::defaults(),
+        EMPTY_QUERY_PLAN,
+        {},
+        {},
+        false,
+        std::nullopt,
+        nullptr});
+  }
+};
+
 class HybridPlanNode : public PlanNode {
  public:
   HybridPlanNode(
       const PlanNodeId& id,
       const RowTypePtr& outputType,
-      const std::shared_ptr<RelAlgExecutionUnit> relAlgExecUnit,
-      const std::shared_ptr<NodeProperty> nodeProperty,
+      const std::shared_ptr<CiderParamContext> ciderParamContext,
       std::shared_ptr<const PlanNode> source)
       : PlanNode(id),
         outputType_(outputType),
-        relAlgExecUnit_(relAlgExecUnit),
-        nodeProperty_(nodeProperty),
+        ciderParamContext_(ciderParamContext),
         sources_{source} {}
 
   const RowTypePtr& outputType() const override {
@@ -57,10 +132,13 @@ class HybridPlanNode : public PlanNode {
     return "hybrid";
   }
 
+  const std::shared_ptr<CiderParamContext> getCiderParamContext() const {
+    return ciderParamContext_;
+  }
+
  private:
-  const std::vector<std::shared_ptr<const PlanNode>> sources_;
   const RowTypePtr outputType_;
-  const std::shared_ptr<RelAlgExecutionUnit> relAlgExecUnit_;
-  const std::shared_ptr<NodeProperty> nodeProperty_;
+  const std::shared_ptr<CiderParamContext> ciderParamContext_;
+  const std::vector<std::shared_ptr<const PlanNode>> sources_;
 };
 } // namespace facebook::velox::core
